% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col.R
\name{preprocess_col}
\alias{preprocess_col}
\title{preprocess_col}
\usage{
preprocess_col(
  url = paste0("http://www.catalogueoflife.org/DCA_Export/zip-fixed/", 2019,
    "-annual.zip"),
  output_paths = c(dwc = "2019/dwc_col.tsv.bz2", common = "2019/common_col.tsv.bz2"),
  dir = file.path(tempdir(), "col")
)
}
\arguments{
\item{url}{Original source}

\item{output_paths}{paths where output will be written: must be two paths, one named dwc, one named common}

\item{dir}{working directory for downloads, will be tempdir() by default}
}
\description{
preprocess_col
}
\details{
NOTE: A list of  all snapshots available from: http://www.catalogueoflife.org/DCA_Export/archive.php
}
