% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb.R
\name{preprocess_fb}
\alias{preprocess_fb}
\title{preprocess_fb}
\usage{
preprocess_fb(
  output_paths = c(dwc = "2019/dwc_fb.tsv.bz2", common = "2019/common_fb.tsv.bz2")
)
}
\arguments{
\item{output_paths}{paths where output will be written: must be two paths, one named dwc, one named common}
}
\description{
preprocess_fb
}
