% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbif.R
\name{preprocess_gbif}
\alias{preprocess_gbif}
\title{preprocess_gbif}
\usage{
preprocess_gbif(
  archive,
  output_paths = c(dwc = "2019/dwc_gbif.tsv.bz2", common = "2019/common_gbif.tsv.bz2")
)
}
\arguments{
\item{output_paths}{paths where output will be written: must be two paths, one named dwc, one named common}
}
\description{
preprocess_gbif
}
