% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iucn.R
\name{preprocess_iucn}
\alias{preprocess_iucn}
\title{preprocess_iucn}
\usage{
preprocess_iucn(
  url = "~/Documents/data/redlist-downloads-2019-11-25",
  output_paths = c(dwc = "2019/dwc_iucn.tsv.bz2", common = "2019/common_iucn.tsv.bz2")
)
}
\arguments{
\item{url}{Original source}

\item{output_paths}{paths where output will be written: must be two paths, one named dwc, one named common}
}
\description{
preprocess_iucn
}
\details{
Before you can use this function, you must first manually downoad IUCN data from your own account.
Go to the Redlist website, create a user, login, and set search options to include taxonomy, synoynms, and common names
Bird data is rate-limited -- filter for Passerine birds, & then for all other birds to get it in two separate downloads
Then also download all other vertabrates & invertabrates
Unzip all downloads to a location specified in path

This approach is much faster & more thorough than the API approach taken in `redlist.R`
}
