% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi.R
\name{preprocess_ncbi}
\alias{preprocess_ncbi}
\title{preprocess_ncbi}
\usage{
preprocess_ncbi(
  url = "ftp://ftp.ncbi.nih.gov/pub/taxonomy/taxdmp.zip",
  output_paths = c(dwc = "2019/dwc_ncbi.tsv.bz2", common = "2019/common_ncbi.tsv.bz2")
)
}
\arguments{
\item{url}{Original source}

\item{output_paths}{paths where output will be written: must be two paths, one named dwc, one named common}
}
\description{
preprocess_ncbi
}
