% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ott.R
\name{preprocess_ott}
\alias{preprocess_ott}
\title{preprocess_ott}
\usage{
preprocess_ott(
  url = "http://files.opentreeoflife.org/ott/ott3.2/ott3.2.tgz",
  output_paths = c(dwc = "2019/dwc_ott.tsv.bz2")
)
}
\arguments{
\item{url}{Original source}

\item{output_paths}{paths where output will be written: must be two paths, one named dwc, one named common}
}
\description{
preprocess_ott
}
