% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slb.R
\name{preprocess_slb}
\alias{preprocess_slb}
\title{preprocess_slb}
\usage{
preprocess_slb(
  output_paths = c(dwc = "2019/dwc_slb.tsv.bz2", common = "2019/common_slb.tsv.bz2")
)
}
\arguments{
\item{output_paths}{paths where output will be written: must be two paths, one named dwc, one named common}
}
\description{
preprocess_slb
}
