% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpl.R
\name{preprocess_tpl}
\alias{preprocess_tpl}
\title{preprocess_tpl}
\usage{
preprocess_tpl(output_paths = c(dwc = "2019/dwc_tpl.tsv.bz2"))
}
\arguments{
\item{output_paths}{paths where output will be written: must be two paths, one named dwc, one named common}
}
\description{
preprocess_tpl
}
