% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col.R
\name{preprocess_col}
\alias{preprocess_col}
\title{preprocess_col}
\usage{
preprocess_col(
  archive,
  output_paths = c(dwc = "data/dwc_col.tsv.gz", common = "data/common_col.tsv.gz",
    dwc_parquet = "data/dwc_col.parquet", common_parquet = "data/common_col.parquet")
)
}
\arguments{
\item{output_paths}{paths where output will be written: must be two paths, one named dwc, one named common}

\item{url}{Source of Catalogue of Life}

\item{dir}{working directory for downloads, will be tempdir() by default}
}
\description{
preprocess_col
}
\details{
NOTE: A list of  all snapshots available from: http://www.catalogueoflife.org/DCA_Export/archive.php
}
