% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itis.R
\name{preprocess_itis}
\alias{preprocess_itis}
\title{preprocess_itis}
\usage{
preprocess_itis(
  archive,
  output_paths = c(dwc = "2021/dwc_itis.tsv.bz2", common = "2021/common_itis.tsv.bz2")
)
}
\arguments{
\item{output_paths}{paths where output will be written: must be two paths, one named dwc, one named common}
}
\description{
preprocess_itis
}
